+++ lib/route_observer_mixin.dart
@@ -1,4 +1,5 @@
 library route_observer_mixin;
 
+export 'src/route_aware_provider.dart';
 export 'src/route_observer_mixin.dart';
 export 'src/route_observer_provider.dart';