+++ lib/src/route_observer_mixin.dart
@@ -1,5 +1,9 @@
+import 'dart:async';
+
+import 'package:disposable_provider/disposable_provider.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
+import 'package:provider/single_child_widget.dart';
 
 mixin RouteObserverMixin<T extends StatefulWidget> on State<T>, RouteAware {
   RouteObserver<ModalRoute> _routeObserver;